+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - Initial module structure
 - Manage the users and service users of a netbird account with the `netbird::User`
   resource
+- Manage the setup keys of a netbird account with the `netbird::SetupKey` resource